Removed template classes and added dummy test
