--- src/main/resources/templates/consultora-nueva.hbs
@@ -1,19 +1,0 @@
-{{# partial "contenido" }}  
-
-<h1>Nueva Consultora</h1>
-
-<form action="/consultoras" method="POST">
-
-  <label for="nombre">Nombre</label>
-	<input type="text" name="nombre">
-	
-	<label for="cantidadEmpleados">Cantidad de Empleados</label>
-	<input type="number" name="cantidadEmpleados">
-	
-	<input type="submit">
-</form>
-
-{{/partial}}
-
-
-{{> layout.hbs }}
