--- src/main/resources/templates/consultora.hbs
@@ -1,1 +1,0 @@
- aca iria el detalle de la consultora {{nombre}}
