+++ src/main/resources/templates/layout.hbs
@@ -12,12 +12,10 @@
   <nav class="navbar navbar-default">
     <div class="container">
       <div class="navbar-header">
-        <a class="navbar-brand"href="/">Consultorea.me</a>
+        <a class="navbar-brand" href="/">Super Gol</a>
       </div>
       <div>
         <ul class="nav navbar-nav">
-          <li><a href="/consultoras">Consultoras</a></li>
-          <li><a href="/proyectos">Proyectos</a></li>
         </ul>
       </div>
     </div>
@@ -27,10 +25,10 @@
   {{# block "contenido" }}
   {{/block}}
   </div>
-  
+
   <footer class="footer">
     <p class="text-center">
-      &copy; 2015 Consultoreame.com
+      &copy; 2015 - Desarrollo de aplicaciones - Aloi y Alvarez
     </p>
   </footer>
   </body>
