+++ src/test/java/unq/dapp/supergol/model/DummyTest.java
@@ -1,0 +1,12 @@
+package unq.dapp.supergol.model;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class DummyTest {
+  @Test
+  public void dummy() {
+    assertTrue(true);
+  }
+}
